Re-organize various scripting editors
