+++ gui/src/app/Scripting/InterpreterTypes.ts
@@ -1,0 +1,18 @@
+export type InterpreterStatus =
+  | "idle"
+  | "loading"
+  | "installing"
+  | "running"
+  | "completed"
+  | "failed";
+
+export const isInterpreterStatus = (x: any): x is InterpreterStatus => {
+  return [
+    "idle",
+    "loading",
+    "installing",
+    "running",
+    "completed",
+    "failed",
+  ].includes(x);
+};
