+++ gui/src/app/Scripting/pyodideWorker/sp_patch_matplotlib.py
